+++ include/Point.h
@@ -8,7 +8,7 @@ class Point: public Tuple
 public:
   Point()=default;
   Point(const Tuple& a) : Tuple(a) {}
-  Point(float x, float y, float z): Tuple(x, y, z, 1.0)
+  Point(double x, double y, double z): Tuple(x, y, z, 1.0)
   {}
 };
